+++ backend/Goalz/Goalz.Application/DTOs/CreateZoneDto.cs
@@ -2,11 +2,10 @@ using NetTopologySuite.Geometries;
 
 namespace Goalz.Core.DTOs
 {
-    public class ZoneDto
+    public class CreateZoneDto
     {
-        public long Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public string Color { get; set; } = string.Empty;
+        public string Color { get; set; } = "#33A661";
         public long? BoundaryId { get; set; }
         public Geometry Boundary { get; set; } = null!;
     }