+++ backend/Goalz/Goalz.Application/DTOs/UpdateZoneDto.cs
@@ -2,12 +2,11 @@ using NetTopologySuite.Geometries;
 
 namespace Goalz.Core.DTOs
 {
-    public class ZoneDto
+    public class UpdateZoneDto
     {
-        public long Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public string Color { get; set; } = string.Empty;
+        public string Color { get; set; } = "#33A661";
         public long? BoundaryId { get; set; }
-        public Geometry Boundary { get; set; } = null!;
+        public Geometry? Boundary { get; set; }
     }
 }